# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContactSubmissionPackageInterest =
  (typeof ContactSubmissionPackageInterest)[keyof typeof ContactSubmissionPackageInterest];

export const ContactSubmissionPackageInterest = {
  starter: "starter",
  pro: "pro",
  authority: "authority",
  general: "general",
} as const;

export interface ContactSubmission {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 2 */
  businessName: string;
  /** @minLength 10 */
  message: string;
  packageInterest?: ContactSubmissionPackageInterest;
}

export interface ContactSubmissionResponse {
  id: number;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}